add implementation of XHR override, move functions to token-util
